Fix build according to Theia refactoring (#259)
* Update according to Theia refactoring
