+++ extensions/che-theia-hosted-plugin-manager-extension/package.json
@@ -10,6 +10,7 @@
   ],
   "dependencies": {
     "@theia/core": "next",
+    "@theia/plugin-dev": "next",
     "@theia/plugin-ext": "next",
     "@eclipse-che/workspace-client": "latest",
     "@eclipse-che/api": "latest"
