+++ extensions/che-theia-hosted-plugin-manager-extension/src/node/che-workspace-hosted-plugin-uri-post-processor.ts
@@ -10,7 +10,7 @@
 
 import { injectable } from 'inversify';
 import URI from '@theia/core/lib/common/uri';
-import { HostedPluginUriPostProcessor } from '@theia/plugin-ext';
+import { HostedPluginUriPostProcessor } from '@theia/plugin-dev';
 import WorkspaceClient, { IRemoteAPI, IRestAPIConfig } from '@eclipse-che/workspace-client';
 import { che } from '@eclipse-che/api';
