+++ extensions/che-theia-hosted-plugin-manager-extension/src/node/hosted-plugin-manager-contribution-extension-backend-module.ts
@@ -9,7 +9,7 @@
  **********************************************************************/
 
 import { ContainerModule } from 'inversify';
-import { HostedPluginUriPostProcessorSymbolName } from '@theia/plugin-ext';
+import { HostedPluginUriPostProcessorSymbolName } from '@theia/plugin-dev';
 import { CheWorkspaceHostedPluginUriPostProcessor } from './che-workspace-hosted-plugin-uri-post-processor';
 
 export default new ContainerModule(bind => {
